Refactoring with moving common models into one project. Removing some dependencies
