+++ devpodcasts.common/Constants/EpisodeConstants.cs
@@ -1,0 +1,17 @@
+﻿namespace devpodcasts.common.Constants;
+
+public class EpisodeConstants
+{
+    public const string TitleElementName = "title";
+    public const string LinkElementName = "link";
+    public const string CategoryElementName = "category";
+    public const string SummaryElementName = "summary";
+    public const string PublishedDateElementName = "pubDate";
+    public const string DurationElementName = "duration";
+    public const string DescriptionElementName = "description";
+    public const string AuthorElementName = "Author";
+    public const string KeywordsElementName = "keywords";
+    public const string EnclosureElementName = "enclosure";
+    public const string TypeElementName = "type";
+    public const string UrlElementName = "url";
+}
