+++ devpodcasts.common/Interfaces/IUpdater.cs
@@ -1,0 +1,6 @@
+﻿namespace devpodcasts.common.Interfaces;
+
+public interface IUpdater : IDisposable
+{
+    Task UpdateDataAsync();
+}
