+++ devpodcasts.common/JsonObjects/PodcastResult.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
-namespace devpodcasts.Services.Core.JsonObjects
+namespace devpodcasts.common.JsonObjects
 {
     public class PodcastResult
     {
