+++ devpodcasts.server.core/appsettings.Development.json
@@ -31,6 +31,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "PodcastDb": ""
+    "PodcastDb": "Server=(localdb)\\msssqllocaldb;Integrated Security=true;"
   }
 }
