+++ devpodcasts.server.core/appsettings.json
@@ -31,7 +31,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "PodcastDb": ""
+    "PodcastDb": "Server=(localdb)\\msssqllocaldb;Integrated Security=true;"
   }
 }
