+++ devpodcasts.services.core/ServiceRunner.cs
@@ -1,4 +1,5 @@
-﻿using devpodcasts.Services.Core.Interfaces;
+﻿using devpodcasts.common.Extensions;
+using devpodcasts.common.Interfaces;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Threading.Tasks;
